pyproject.toml
[project]
name = "copier"
version = "0.1.0"
description = "Programmatically copy file to and from configured locations"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "argparse>=1.4.0",
    "pathlib>=1.0.1",
    "pyyaml>=6.0.3",
]


classifiers = ["Private :: Do Not Upload"]
authors = [
    { name = "Alvise Bruniera", email = "alvisebruniera@gmail.com" },
    { name = "Lucide", email = "33134209+Lucide@users.noreply.github.com" },
]

[project.scripts]
copier = "copier:main"

[build-system]
requires = ["uv_build>=0.8.5,<0.9.0"]
build-backend = "uv_build"
